Move get_definition function to a separate module, clean up imports
